setting 모듈 main 모듈 내부로 이동
main 모듈 내에서 setting 모듈 기능을 구현하는 것 으로 변경
