+++ main/build.gradle.kts
@@ -19,6 +19,7 @@ android {
 
   buildFeatures {
     viewBinding = true
+    buildConfig = true
   }
 }
