+++ main/src/main/kotlin/us/wedemy/eggeum/android/main/ui/home/HomeFragment.kt
@@ -5,7 +5,7 @@
  * Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
  */
 
-package us.wedemy.eggeum.android.main.ui
+package us.wedemy.eggeum.android.main.ui.home
 
 import android.os.Bundle
 import android.view.View
