+++ main/src/main/kotlin/us/wedemy/eggeum/android/main/ui/myaccount/SettingFragment.kt
@@ -5,13 +5,13 @@
  * Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
  */
 
-package us.wedemy.eggeum.android.main.ui
+package us.wedemy.eggeum.android.main.ui.myaccount
 
 import dagger.hilt.android.AndroidEntryPoint
 import us.wedemy.eggeum.android.common.ui.BaseFragment
-import us.wedemy.eggeum.android.main.databinding.FragmentCafeInfoBinding
+import us.wedemy.eggeum.android.main.databinding.FragmentSettingBinding
 
 @AndroidEntryPoint
-class CafeInfoFragment : BaseFragment<FragmentCafeInfoBinding>() {
-  override fun getViewBinding() = FragmentCafeInfoBinding.inflate(layoutInflater)
+class SettingFragment : BaseFragment<FragmentSettingBinding>() {
+  override fun getViewBinding() = FragmentSettingBinding.inflate(layoutInflater)
 }
