+++ main/src/main/kotlin/us/wedemy/eggeum/android/main/ui/search/CafeDetailFragment.kt
@@ -5,7 +5,14 @@
  * Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
  */
 
-package us.wedemy.eggeum.android.main.ui
+/*
+ * Designed and developed by Wedemy 2023.
+ *
+ * Licensed under the MIT.
+ * Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
+ */
+
+package us.wedemy.eggeum.android.main.ui.search
 
 import android.os.Bundle
 import android.view.View
