+++ main/src/main/kotlin/us/wedemy/eggeum/android/main/ui/search/CafeImageFragment.kt
@@ -5,25 +5,29 @@
  * Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
  */
 
-package us.wedemy.eggeum.android.main.ui
+package us.wedemy.eggeum.android.main.ui.search
 
 import android.os.Bundle
 import android.view.View
 import dagger.hilt.android.AndroidEntryPoint
 import us.wedemy.eggeum.android.common.ui.BaseFragment
-import us.wedemy.eggeum.android.common.extension.addDivider
-import us.wedemy.eggeum.android.design.R
-import us.wedemy.eggeum.android.main.databinding.FragmentSearchCafeBinding
+import us.wedemy.eggeum.android.main.databinding.FragmentCafeImageBinding
+import us.wedemy.eggeum.android.main.ui.adapter.CafeImageAdapter
 
 @AndroidEntryPoint
-class SearchCafeFragment : BaseFragment<FragmentSearchCafeBinding>() {
-  override fun getViewBinding() = FragmentSearchCafeBinding.inflate(layoutInflater)
+class CafeImageFragment : BaseFragment<FragmentCafeImageBinding>() {
+
+  private val cafeImageAdapter by lazy {
+    CafeImageAdapter { _ -> run {} }
+  }
+
+  override fun getViewBinding() = FragmentCafeImageBinding.inflate(layoutInflater)
 
   override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
     super.onViewCreated(view, savedInstanceState)
-    binding.rvSearchCafe.apply {
+    binding.rvCafeImage.apply {
       setHasFixedSize(true)
-      addDivider(R.color.gray_300)
+      adapter = cafeImageAdapter
     }
   }
 }
