+++ main/src/main/kotlin/us/wedemy/eggeum/android/main/viewmodel/EditMyInfoViewModel.kt
@@ -5,7 +5,7 @@
  * Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
  */
 
-package us.wedemy.eggeum.android.setting.viewmodel
+package us.wedemy.eggeum.android.main.viewmodel
 
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
