--- setting/src/main/res/layout/activity_setting.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Designed and developed by Wedemy 2023.
-  ~
-  ~ Licensed under the MIT.
-  ~ Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
-  -->
-
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto"
-  xmlns:tools="http://schemas.android.com/tools"
-  android:layout_width="match_parent"
-  android:layout_height="match_parent"
-  tools:context=".ui.SettingActivity">
-
-  <androidx.fragment.app.FragmentContainerView
-    android:id="@+id/nav_host_fragment"
-    android:name="androidx.navigation.fragment.NavHostFragment"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    app:defaultNavHost="true"
-    app:navGraph="@navigation/nav_setting" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
