--- setting/src/main/res/values/strings.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Designed and developed by Wedemy 2023.
-  ~
-  ~ Licensed under the MIT.
-  ~ Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
-  -->
-
-<resources>
-  <string name="edit_my_info">내 정보 수정</string>
-  <string name="nickname">닉네임 <font color="#14b8a6">*</font></string>
-  <string name="nickname_hint">닉네임</string>
-  <string name="email">이메일</string>
-  <string name="edit">수정하기</string>
-  <string name="temp">최소 2자 이상 입력해주세요.</string>
-  <string name="empty_text_error">닉네임을 입력해주세요.</string>
-  <string name="too_short_text_error">최소 2자 이상 입력해주세요.</string>
-  <string name="default_email">eggeum@gamil.com</string>
-</resources>
