--- setting/src/main/res/values/themes.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Designed and developed by Wedemy 2023.
-  ~
-  ~ Licensed under the MIT.
-  ~ Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
-  -->
-
-<resources>
-
-  <style name="Theme.Eggeum.Setting" parent="Theme.MaterialComponents.Light.NoActionBar" />
-</resources>
